Revert "ファイルを移動"
This reverts commit d0249de43f6f70fe5a4c635dbc2af63fe3e8297d.

+++ Template/Vector2.h
@@ -63,7 +63,7 @@ public:
 
 	//長さを返す関数
 	float Magnitude() {
-		return (float)sqrt(x * x + y * y);
+		return sqrt(x * x + y * y);
 	}
 
 	//単位ベクトルを返す関数
